Create visualsnippet
You have a textbox where can be put numbers only, not letters or symbols. This little piece of code can be useful to make an exception, and i have used it so many times that i want to add it here.
The word "Errore" is the italian word for "Error", i'm italian.